Split ManagerAgent into focused modules (473 → 156 lines)
The orchestrator had five tangled concerns: tier check + dispatch
(real role), config caching, agent selection (three classifiers),
research-brief formatting, and a daily routing log. Each now lives
in its own file under orchestrator/manager-agent/:

- types.ts             OrchestratorRequest / OrchestratorResponse /
                       AgentName / HubConfig
- config.ts            loadHubConfig + flushHubConfig — the cached
                       four-YAML view; cache flush comment now
                       explains why empty catch is safe
- agent-selection.ts   selectAgent + isWebLookupIntent +
                       classifyWithLlm + classifyWithKeywords. The
                       three keyword arrays (finance / business /
                       operations) are top-level constants instead
                       of inline-in-function so the linter can
                       hoist them and the next reader can spot them.
- format.ts            formatBrief — research-brief markdown
- routing-log.ts       logRouting — daily append-only audit trail

ManagerAgent.ts keeps `route()` and `flushConfig()` as the only
static methods. External callers (route handlers, ruleset writes)
don't change — same class, same method names.
